chore(dynamic-catalog): move resolveCatalogLookup
