+++ packages/ui/src/components/Visualization/Canvas/Form/fields/EndpointField/EndpointField.test.tsx
@@ -28,7 +28,6 @@ describe('EndpointField', () => {
     getCatalog: jest.fn(),
     setCatalog: jest.fn(),
     clearRegistry: jest.fn(),
-    resolveCatalogLookup: jest.fn(),
   };
 
   const createWrapper = (
