+++ packages/ui/src/dynamic-catalog/catalog-modal.provider.test.tsx
@@ -21,7 +21,6 @@ describe('CatalogModalProvider', () => {
     getCatalog: jest.fn(),
     setCatalog: jest.fn(),
     clearRegistry: jest.fn(),
-    resolveCatalogLookup: jest.fn(),
   };
 
   const createWrapper = (
