+++ packages/ui/src/dynamic-catalog/dynamic-catalog-registry.test.ts
@@ -358,89 +358,6 @@ describe('DynamicCatalogRegistry', () => {
     });
   });
 
-  describe('resolveCatalogLookup', () => {
-    const timerComponent = {
-      component: { name: 'timer', syntax: 'timer:name' },
-      properties: {},
-      propertiesSchema: {},
-    } as ICamelComponentDefinition;
-
-    const kameletComponent = {
-      component: { name: 'kamelet' },
-      properties: {},
-      propertiesSchema: {},
-    } as ICamelComponentDefinition;
-
-    const chuckNorrisKamelet = {
-      kind: 'Kamelet',
-      metadata: { name: 'chuck-norris-source' },
-      spec: { definition: {} },
-    } as IKameletDefinition;
-
-    beforeEach(() => {
-      const componentCatalog = new DynamicCatalog<ICamelComponentDefinition>({
-        id: 'component-provider',
-        fetch: (key) =>
-          Promise.resolve(
-            ({ timer: timerComponent, kamelet: kameletComponent } as Record<string, ICamelComponentDefinition>)[key],
-          ),
-        fetchAll: () => Promise.resolve({ timer: timerComponent, kamelet: kameletComponent }),
-      });
-
-      const kameletCatalog = new DynamicCatalog<IKameletDefinition>({
-        id: 'kamelet-provider',
-        fetch: (key) =>
-          Promise.resolve(({ 'chuck-norris-source': chuckNorrisKamelet } as Record<string, IKameletDefinition>)[key]),
-        fetchAll: () => Promise.resolve({ 'chuck-norris-source': chuckNorrisKamelet }),
-      });
-
-      registry.setCatalog(CatalogKind.Component, componentCatalog);
-      registry.setCatalog(CatalogKind.Kamelet, kameletCatalog);
-    });
-
-    it('should return undefined for an empty component name', async () => {
-      const lookup = await registry.resolveCatalogLookup('');
-
-      expect(lookup).toBeUndefined();
-    });
-
-    it('should return a component from the catalog lookup', async () => {
-      const lookup = await registry.resolveCatalogLookup('timer');
-
-      expect(lookup).toEqual({
-        catalogKind: CatalogKind.Component,
-        definition: timerComponent,
-      });
-    });
-
-    it('should return a kamelet from the catalog lookup', async () => {
-      const lookup = await registry.resolveCatalogLookup('kamelet:chuck-norris-source');
-
-      expect(lookup).toEqual({
-        catalogKind: CatalogKind.Kamelet,
-        definition: chuckNorrisKamelet,
-      });
-    });
-
-    it('should return the kamelet component for unknown kamelets', async () => {
-      const lookup = await registry.resolveCatalogLookup('kamelet:non-existing-kamelet');
-
-      expect(lookup).toEqual({
-        catalogKind: CatalogKind.Component,
-        definition: kameletComponent,
-      });
-    });
-
-    it('should pass forceFresh to underlying catalog lookups', async () => {
-      const componentCatalog = registry.getCatalog(CatalogKind.Component)!;
-      const getSpy = jest.spyOn(componentCatalog, 'get');
-
-      await registry.resolveCatalogLookup('timer', { forceFresh: true });
-
-      expect(getSpy).toHaveBeenCalledWith('timer', { forceFresh: true });
-    });
-  });
-
   it('should maintain a singleton registry', () => {
     const registry1 = DynamicCatalogRegistry.get();
     const registry2 = DynamicCatalogRegistry.get();
